Extract functions that interact with kubectl to another module
